# clients/typescript/scripts/openapi.snapshot.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Spatial.Host | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5299/"
    }
  ],
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "AuthEndpoints"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/maps/{name}": {
      "get": {
        "tags": [
          "AdminEndpoints"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Map"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthTokenResponse": {
        "required": [
          "token",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "LoginRequest": {
        "required": [
          "username",
          "password"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "Map": {
        "required": [
          "name",
          "store",
          "layers",
          "services"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "store": {
            "type": "string"
          },
          "layers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MapLayer"
            }
          },
          "services": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MapServiceKind"
            }
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "copyright": {
            "type": [
              "null",
              "string"
            ]
          },
          "metadataXml": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "MapLayer": {
        "required": [
          "dataset",
          "layerId"
        ],
        "type": "object",
        "properties": {
          "dataset": {
            "type": "string"
          },
          "layerId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "style": {
            "type": [
              "null",
              "string"
            ]
          },
          "kind": {
            "$ref": "#/components/schemas/MapLayerKind"
          },
          "store": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "MapLayerKind": {
        "enum": [
          "feature",
          "image"
        ],
        "default": "feature"
      },
...
      "MapServiceKind": {
        "enum": [
          "feature",
          "map",
          "tiles",
          "wms",
          "wfs",
          "image"
        ]
      },
...
    }
  },
...
}